HACK: use static libraries by default
This probably doesn't matter that much, but to allow debugging directly
from GLEW, we want to use static linking to avoid having to copy any
DLLs around.

Sadly, I can't find a directed way of doing this, but linking statically
for all libraries can't be that bad; we don't really have any, and if we
get some, why would we want them shared?