leak all the imported textures
This is not the time to fix structural issues. We need to keep textures
longer than the current scope in some cases, so let's allow that. Yeah,
it's ugly. We can probably use some unique_ptr magic or something to fix
it, but now is not the time. Let's just heap-allocate and leak
everything. The OS will clean it up on shutdown anyway...